nominal_polyumi/requirements.txt
foxglove-sdk
numpy
websockets==13.1
mcap
mcap-ros2-support
Pillow

-e ../ros2_ws/src/polyumi_pi_msgs